+++ src/sitecore-rocks/src/data/FieldUri.ts
@@ -1,8 +1,5 @@
-import { DatabaseUri } from "./DatabaseUri";
 import { isString } from "./helpers";
-import { ItemUri } from "./ItemUri";
-import { ItemVersionUri } from "./ItemVersionUri";
-import { WebsiteUri } from "./WebsiteUri";
+import { DatabaseUri, ItemUri, ItemVersionUri, WebsiteUri } from "./index";
 
 export class FieldUri {
     public static readonly empty = new FieldUri(ItemVersionUri.empty, "{00000000-0000-0000-0000-000000000000}");