+++ src/sitecore-rocks/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Webservice version check
 
 ## [0.0.4] - 2017-10-04
 ### Added